[refactor] modify based on feedback(1) - rand -> random - add util package - block comment -> docx comment - constantization dummy string - use assertEquals
