+++ src/main/java/racingcar/RacingCarGame.java
@@ -3,7 +3,7 @@ package racingcar;
 import java.util.List;
 
 public class RacingCarGame {
-    /*
+    /**
      * Racing Game을 실제로 구동시키는 클래스입니다.
      * 구현한 모든 클래스가 조합되어 사용되는 부분
      */
