+++ src/test/java/racingcar/CarListTest.java
@@ -7,13 +7,15 @@ import java.util.List;
 
 public class CarListTest {
 
+    private final static String CAR_DUMMY_STRING = "dummy";
+
     @Test
     void duplicatedNameException() {
         CarList carList = new CarList();
         Assertions.assertThatExceptionOfType(RuntimeException.class)
                 .isThrownBy(() -> {
-                    carList.add(new Car("lion"));
-                    carList.add(new Car("lion"));
+                    carList.add(new Car(CAR_DUMMY_STRING));
+                    carList.add(new Car(CAR_DUMMY_STRING));
                 });
     }
