+++ src/test/java/racingcar/util/RandomUtilTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
@@ -6,8 +6,7 @@ import org.junit.jupiter.api.RepeatedTest;
 public class RandomUtilTest {
     @RepeatedTest(100)
     void generateRandomZeroToNine() {
-        RandomUtil randomUtil = new RandomUtil();
-        int randomNumber = randomUtil.generateRandom();
+        int randomNumber = RandomUtil.generateRandom();
         Assertions.assertThat(randomNumber)
                 .isBetween(0, 9);
     }
